Update the ChangeLog and add darkstar.nuthouse.org to the README.
Stop appending a space to $CONFIG.TABKEY_NICKCOMP_SUFFIX when a nickname
is successfully completed. Instead, the space was added to the actual value
of the TABKEY_NICKCOMP_SUFFIX config variable.

Copied the values of the MODLIST_* format variables in the darkstar theme
to the defaults located in darkstar.irc.

Also added a small "Loading DarkStar <version>" msg in darkstar.irc.