misc.dsm 0.7.4:
The raw_irc hook created on startup if $CONFIG.IGNORE_CTCP_FINGER is true
had 'A' instead of '^A' causing it to miss FINGER requests.

Rather than simply fix this mistake I decided to rewrite the way this
feature works. Now, instead of adding and removing the raw_irc hook
whenever it is needed, a raw_irc hook with the ambiguous mode ('?') will
be added at load time (like everything else) and use a conditional to
determine whether the client should continue processing the request.

This is a much cleaner way to do things and should have less potential
to cause problems.