Added a new completion procedure called mask and 3 new completion rules, 2 of which use the new mask procedure.
The mask procedure attempts to convert the last word on the input prompt
(hopefully a nick) into a hostmask suitable for use with commands like
/ignore and /ban. It will first try to get the userhost from the client's
cache and, failing that, will query the server.

The new rules add the hostmask completion described above for the /ignore
and /ban commands as well as completion for the various ignore types.