Simple changes to /LOADMOD and /UNLOADMOD that should prevent the possibility of modules causing trouble with the input_prompt when $CONFIG.LOADMODULE_VERBOSE is true.
The value of LOADMODULE_VERBOSE is now stored in a local variable at
the beginning of the aliases and this local variable is used in all
conditionals throughout the rest of the aliases. This should prevent
modules from being able to change the value on us before the alias
finishes executing.