Fix a problem reported by tgz in which the input prompt would not get reset after loading/unloading modules when $CONFIG.LOADMODULE_VERBOSE is false.
This bug was only evident when using the scrollz theme due to the fact that
it had an empty value for INPUT_PROMPT which caused the attempt to reset
the prompt to fail because of the missing argument. The loader now checks
for an empty value in $oldprompt and unsets INPUT_PROMPT accordingly.