Fixes for EPIC4-1.1.7.
These are mostly changes to the key bindings in window.dsm. Special thanks
to Jeremy Nelson for doing the majority of the work involved in this.

I have also updated the client requirements to 1.1.7 in light of the recent
changes to /defer which could potentially create double expansion backdoors
if newer versions of DarkStar are used with older versions of EPIC4.

If no problems are found this will likely become 0.2.8.